requirements.txt
discord.py
tpblite
requests
seedrcc
python-dotenv
